[Chat] Move InMemory store from Bridge to Chat root
