+++ src/ai-bundle/tests/Profiler/DataCollectorTest.php
@@ -17,8 +17,8 @@ use Symfony\AI\AiBundle\Profiler\DataCollector;
 use Symfony\AI\AiBundle\Profiler\TraceableChat;
 use Symfony\AI\AiBundle\Profiler\TraceableMessageStore;
 use Symfony\AI\AiBundle\Profiler\TraceablePlatform;
-use Symfony\AI\Chat\Bridge\Memory\Store as InMemoryStore;
 use Symfony\AI\Chat\Chat;
+use Symfony\AI\Chat\InMemory\Store as InMemoryStore;
 use Symfony\AI\Platform\Message\Content\Text;
 use Symfony\AI\Platform\Message\Message;
 use Symfony\AI\Platform\Message\MessageBag;
