+++ src/ai-bundle/tests/Profiler/TraceableChatTest.php
@@ -14,8 +14,8 @@ namespace Symfony\AI\AiBundle\Tests\Profiler;
 use PHPUnit\Framework\TestCase;
 use Symfony\AI\Agent\AgentInterface;
 use Symfony\AI\AiBundle\Profiler\TraceableChat;
-use Symfony\AI\Chat\Bridge\Memory\Store as InMemoryStore;
 use Symfony\AI\Chat\Chat;
+use Symfony\AI\Chat\InMemory\Store as InMemoryStore;
 use Symfony\AI\Platform\Message\Message;
 use Symfony\AI\Platform\Message\MessageBag;
 use Symfony\AI\Platform\Message\UserMessage;
